# speed up 'tag' filter rule SQL query