Instead of doing a JOIN and comparing the tag name for every instance of a tag in the library, extract just the tag IDs for tags with matching names and perform a set membership to select the images of interest.

On a library with 219k images:
-  tag: 'not tagged'  goes from 0.65 to 0.50 seconds
-  a simple match (with or without case folding) for a tag with 41k instances goes from 0.32 to 0.08 seconds
-    .....  3k instances goes from 0.08 to 0.01 seconds
-  a subtree match including 90k images goes from 0.32 to 0.24 seconds

This is a companion to #9017 and #9028, but can be merged independently of either.